+++ LPC/Cepstrum_and_Spectrum.cpp
@@ -1,6 +1,6 @@
 /* Cepstrum_and_Spectrum.cpp
  *
- * Copyright (C) 1994-2012, 2015 David Weenink
+ * Copyright (C) 1994-2017 David Weenink
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -37,7 +37,7 @@ static autoCepstrum Spectrum_to_Cepstrum_cmplx (Spectrum me) {
 
 		// Copy magnitude-squared and unwrapped phase.
 
-		for (long i = 1; i <= my nx; i ++) {
+		for (integer i = 1; i <= my nx; i ++) {
 			double xa = unwrap -> z[1][i];
 			sx -> z[1][i] = xa > 0.0 ? 0.5 * log (xa) : -300.0;
 			sx -> z[2][i] = unwrap -> z[2][i];
@@ -59,15 +59,15 @@ autoPowerCepstrum Spectrum_to_PowerCepstrum (Spectrum me) {
 	try {
 		autoSpectrum dBspectrum = Data_copy (me);
 		double *re = dBspectrum -> z[1], *im = dBspectrum -> z[2];
-		for (long i = 1; i <= dBspectrum -> nx; i ++) {
-			re[i] = log (re[i] * re[i] + im[i] * im[i] + 1e-300);
-			im[i] = 0.0;
+		for (integer i = 1; i <= dBspectrum -> nx; i ++) {
+			re [i] = log (re [i] * re [i] + im [i] * im [i] + 1e-300);
+			im [i] = 0.0;
 		}
 		autoSound cepstrum = Spectrum_to_Sound (dBspectrum.get());
 		autoPowerCepstrum thee = PowerCepstrum_create (0.5 / my dx, my nx);
-		for (long i = 1; i <= thy nx; i++) {
-			double val = cepstrum -> z[1][i];
-			thy z[1][i] = val * val;
+		for (integer i = 1; i <= thy nx; i ++) {
+			double val = cepstrum -> z [1] [i];
+			thy z [1] [i] = val * val;
 		}
 		return thee;
 	} catch (MelderError) {
@@ -79,15 +79,15 @@ autoCepstrum Spectrum_to_Cepstrum (Spectrum me) {
 	try {
 		autoSpectrum dBspectrum = Data_copy (me);
 		double *re = dBspectrum -> z[1], *im = dBspectrum -> z[2];
-		for (long i = 1; i <= dBspectrum -> nx; i++) {
-			re[i] = log (re[i] * re[i] + im[i] * im[i] + 1e-300);
-			im[i] = 0.0;
+		for (integer i = 1; i <= dBspectrum -> nx; i ++) {
+			re [i] = log (re [i] * re [i] + im [i] * im [i] + 1e-300);
+			im [i] = 0.0;
 		}
 		autoSound cepstrum = Spectrum_to_Sound (dBspectrum.get());
 		autoCepstrum thee = Cepstrum_create (0.5 / my dx, my nx);
-		for (long i = 1; i <= thy nx; i++) {
-			double val = cepstrum -> z[1][i];
-			thy z[1][i] = val;
+		for (integer i = 1; i <= thy nx; i ++) {
+			double val = cepstrum -> z [1] [i];
+			thy z [1] [i] = val;
 		}
 		return thee;
 	} catch (MelderError) {
@@ -98,16 +98,16 @@ autoCepstrum Spectrum_to_Cepstrum (Spectrum me) {
 autoSpectrum Cepstrum_to_Spectrum (Cepstrum me) { //TODO power cepstrum
 	try {
 		autoCepstrum cepstrum = Data_copy (me);
-		cepstrum ->  z[1][1] = my z[1][1];
-		for (long i = 2; i <= cepstrum -> nx; i++) {
-			cepstrum -> z[1][i] = 2 * my z[1][i];
+		cepstrum ->  z [1] [1] = my z [1] [1];
+		for (integer i = 2; i <= cepstrum -> nx; i ++) {
+			cepstrum -> z [1] [i] = 2 * my z [1] [i];
 		}
 		autoSpectrum thee = Sound_to_Spectrum ((Sound) cepstrum.get(), 1);
 
-		double *re = thy z[1], *im = thy z[2];
-		for (long i = 1; i <= thy nx; i ++) {
-			re[i] =  exp (0.5 * re[i]);   // i.e., sqrt (exp(re [i]))
-			im[i] = 0.0;
+		double *re = thy z [1], *im = thy z [2];
+		for (integer i = 1; i <= thy nx; i ++) {
+			re [i] =  exp (0.5 * re [i]);   // i.e., sqrt (exp(re [i]))
+			im [i] = 0.0;
 		}
 		return thee;
 	} catch (MelderError) {
