+++ dwsys/Command.cpp
@@ -1,6 +1,6 @@
 /* Command.cpp
  *
- * Copyright (C) 1994-2011, 2015 David Weenink
+ * Copyright (C) 1994-2017 David Weenink
  *
  * This code is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -56,7 +56,7 @@ Command CommandHistory_getItem (CommandHistory me) {
 }
 
 void CommandHistory_insertItem_move (CommandHistory me, autoCommand command) {
-	for (long i = my size; i >= my current + 1; i --) {
+	for (integer i = my size; i >= my current + 1; i --) {
 		my removeItem (i);
 	}
 	my addItem_move (command.move());
@@ -78,8 +78,8 @@ int CommandHistory_offright (CommandHistory me) {
 	return my size == 0 || my current == my size + 1;
 }
 
-const char32 *CommandHistory_commandName (CommandHistory me, long offsetFromCurrent) {
-	long pos = my current + offsetFromCurrent;
+const char32 *CommandHistory_commandName (CommandHistory me, integer offsetFromCurrent) {
+	integer pos = my current + offsetFromCurrent;
 	return pos >= 1 && pos <= my size ? Thing_getName (my at [pos]) : nullptr;
 }
